Update GO_NOGO_training
Now the code has been updated to be used in freely-moving electrophysiological recordings.
It has been introduced a timer, in which one can set the recording period. Both starting and ending points are reported by a +3.3V TTL pulse from the sync pin, which can be recorded by an amplifier to have TTL timestamps. After ending point the lever cannot be pressed anymore.